Closes #687: Move overlay and progress bar to browser/ pkg.
widget/ is more for re-usable widgets for all over the app, rather than
for a specific feature.
