+++ app/src/main/java/org/mozilla/focus/browser/BrowserNavigationOverlay.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.widget
+package org.mozilla.focus.browser
 
 import android.content.Context
 import android.graphics.drawable.Drawable
@@ -16,6 +16,7 @@ import org.mozilla.focus.R
 import org.mozilla.focus.autocomplete.UrlAutoCompleteFilter
 import org.mozilla.focus.telemetry.TelemetryWrapper
 import org.mozilla.focus.utils.Settings
+import org.mozilla.focus.widget.InlineAutocompleteEditText
 
 enum class NavigationEvent {
     HOME, SETTINGS, BACK, FORWARD, RELOAD, LOAD, TURBO, RELOAD_YT, PIN_ACTION;
